Linux/Mac patcher: Fix executable bit
It isn't marked as an executable so anyone who tries running it just
won't work. This is a common issue for any application made by Windows
developers and linux users would have to chmod +x it. This just applies
it so when someone downloads it from now on this shouldn't be a required
step.